Fix: preserve string[pyarrow] dtype in str.split with expand=True
