+++ dask/dataframe/dask_expr/_str_accessor.py
@@ -182,13 +182,17 @@ class SplitMap(FunctionMap):
     @functools.cached_property
     def _meta(self):
         delimiter = " " if self.pat is None else self.pat
+
         meta = meta_nonempty(self.frame._meta)
-        sample_dtype = self.frame._meta.dtypes
+
+        sample_dtype = self.frame._meta.dtype
+
         meta = self.frame._meta._constructor(
             [delimiter.join(["a"] * (self.n + 1))],
             index=meta.iloc[:1].index,
             dtype=sample_dtype if not self.expand else None,
         )
+
         meta = getattr(meta.str, self.attr)(n=self.n, expand=self.expand, pat=self.pat)
 
         if self.expand:
